Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/geolocation-API/getCurrentPosition_IDL.https-expected.txt

Clear all Geolocation permissions before running this test. If prompted for permission, please allow.


Harness Error (TIMEOUT), message = null

TIMEOUT getCurrentPosition success callback tests Test timed out
PASS getCurrentPosition error callback tests
PASS GeolocationPositionError toString
PASS PERMISSION_DENIED value is 1
PASS POSITION_UNAVAILABLE is 2
PASS TIMEOUT value is 3
